// src/integrations/supabase/types.ts
export type Database = {
  public: {
    Tables: {
// ...
      profiles: {
        Row: {
          avatar_url: string | null
          business_details: string | null
          business_name: string | null
          contact_number: string | null
          full_name: string | null
          id: string
          is_active: boolean
          is_featured: boolean | null
          last_seen: string | null
          updated_at: string
        }
        Insert: {
          avatar_url?: string | null
          business_details?: string | null
          business_name?: string | null
          contact_number?: string | null
          full_name?: string | null
          id: string
          is_active?: boolean
          is_featured?: boolean | null
          last_seen?: string | null
          updated_at?: string
        }
        Update: {
          avatar_url?: string | null
          business_details?: string | null
          business_name?: string | null
          contact_number?: string | null
          full_name?: string | null
          id?: string
          is_active?: boolean
          is_featured?: boolean | null
          last_seen?: string | null
          updated_at?: string
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}